fix: revamp doc
