+++ literalai/api/README.md
@@ -1,0 +1,7 @@
+# Literal AI API
+
+This module contains the APIs to directly interact with the Literal AI platform.
+
+`BaseLiteralAPI` has all the methods prototypes and is the source of truth when it comes to documentation.
+
+Inheriting from `BaseLiteralAPI` are `LiteralAPI` and `AsyncLiteralAPI`, the synchronous and asynchronous APIs respectively.
