+++ literalai/api/helpers/prompt_helpers.py
@@ -6,7 +6,7 @@ from literalai.prompt_engineering.prompt import Prompt, ProviderSettings
 if TYPE_CHECKING:
     from literalai.api import LiteralAPI
 
-from literalai.api import gql
+from literalai.api.helpers import gql
 
 
 def create_prompt_lineage_helper(name: str, description: Optional[str] = None):
@@ -86,7 +86,7 @@ def create_prompt_variant_helper(
         "tools": tools,
     }
 
-    def process_response(response):
+    def process_response(response) -> Optional[str]:
         variant = response["data"]["createPromptExperiment"]
         return variant["id"] if variant else None
